+++ apps/admin/src/features/users/index.ts
@@ -1,2 +1,2 @@
-export { Following } from '../user/following/ui/Following'
+export { Following } from '../user/ui/Following'
 export { UsersTable } from './users-list/ui/users-table/UsersTable'
