--- apps/admin/src/pages/users/[id]/following/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-import { Following } from '@/features/users'
-
-export default function Users() {
-  return <Following />
-}
