+++ apps/admin/src/pages/users/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 
 import { UsersTable } from '@/features/users'
-
 export default function Users() {
   return <UsersTable />
 }
