+++ packages/shared/src/locales/followers/index.ts
